tests: fix signedness compilation warnings

+++ src/test/common/test_safe_io.cc
@@ -18,11 +18,11 @@ TEST(SafeIO, safe_read_file) {
   ASSERT_NE(fd, -1);
   const char buf[] = "0123456789";
   for (int i = 0; i < 8; i++) {
-    ASSERT_EQ(sizeof(buf), write(fd, buf, sizeof(buf)));
+    ASSERT_EQ((ssize_t)sizeof(buf), write(fd, buf, sizeof(buf)));
   }
   ::close(fd);
   char rdata[80];
-  ASSERT_EQ(sizeof(rdata),
+  ASSERT_EQ((int)sizeof(rdata),
 	    safe_read_file(".", fname, rdata, sizeof(rdata)));
   for (char *p = rdata, *end = rdata+sizeof(rdata); p < end; p+=sizeof(buf)) {
     ASSERT_EQ(0, std::memcmp(p, buf, std::min(size_t(end-p), sizeof(buf))));
